Add check for nested translations
Laravel translations can be nested and thus the translator will return
an array for these types of translations.  Passing an array into the
FormBuilder where is it expecting a string results in an error when the
form is rendered due to htmlentities not knowing what to do with an
array.  This change akes sure that the returned translation is actually
a string before using it.

I also updated the test to check for this case.